Add productos en firebase

+++ app/src/main/java/com/example/hosteleriapp/Objetos/Producto.kt
@@ -1,3 +1,3 @@
 package com.example.hosteleriapp.Objetos
 
-data class Producto(val correo:String, val nombre:String, var descripcion:String, var precio:Int)
+data class Producto(val correo:String, val nombre:String, var descripcion:String, var precio: Double)
